+++ main.lua
@@ -2,22 +2,50 @@ local playerGUID = UnitGUID("player")
 local lastCritTime = time()
 
 CritWow = {
-
-}
-
-local defaults = {
-    playsSound = true,
-    debounce = 1,
+    modes = {
+        wowenWilson = {
+            id = "wowenWilson",
+            name = "Wowen Wilson",
+            count = 19,
+            soundPath = "Interface\\AddOns\\CritWow\\sounds\\wowen_wilson\\"
+        },
+        dnc = {
+            id = "dnc",
+            name = "DNC",
+            dncChance = 92,
+            soundPath = "Interface\\AddOns\\CritWow\\sounds\\dnc\\"
+        }
+    },
+    defaults = {
+        playsSound = true,
+        debounce = 1,
+        mode = "wowenWilson",
+    }
 }
 
 function CritWow:PlaySound()
-    PlaySoundFile("Interface\\AddOns\\CritWow\\sounds\\" .. tostring(math.random(1, 17)) .. ".mp3", "master")
+    local path = self.modes.wowenWilson.soundPath
+    local fileName = ""
+
+    if CritWowDB.mode == "dnc" then
+        path = self.modes.dnc.soundPath
+
+        if math.random(1, 100) <= self.modes.dnc.dncChance then
+            fileName = "dnc.mp3"
+        else
+            fileName = "igc.mp3"
+        end
+    else
+        fileName = tostring(math.random(1, self.modes.wowenWilson.count)) .. ".mp3"
+    end
+
+    PlaySoundFile(path .. fileName, "master")
 end
 
 function CritWow:SetDebounceSliderText()
     local debounceTitle = "Unlimited"
-    if CritWowDB.debounce > 0 then
-        debounceTitle = "Once per " .. CritWowDB.debounce .. "s"
+    if self.db.debounce > 0 then
+        debounceTitle = "Once per " .. self.db.debounce .. "s"
     end
 
     getglobal("CritWowDebounceSpeedSlider").Text:SetText("Wows: " .. debounceTitle);
@@ -28,24 +56,36 @@ f:RegisterEvent("ADDON_LOADED")
 f:RegisterEvent("COMBAT_LOG_EVENT_UNFILTERED")
 f:SetScript("OnEvent", function(self, event, addOnName)
     if event == "ADDON_LOADED" and addOnName == "CritWow" then
-        self:SetDefaults()
         self:InitializeOptions()
     end
     self:COMBAT_LOG_EVENT_UNFILTERED(CombatLogGetCurrentEventInfo())
 end)
 
-function f:SetDefaults()
-    CritWowDB = CritWowDB or CopyTable(defaults)
+function CritWow:SetDefaults()
+    CritWowDB = CritWowDB or CopyTable(self.defaults)
+
+    -- Set any missing config options
+    for idx, val in pairs(self.defaults) do
+        if not CritWowDB[idx] then
+            CritWowDB[idx] = self.defaults[idx]
+        end
+    end
+
+    self.db = CritWowDB
 end
 
 function f:InitializeOptions()
+    CritWow:SetDefaults()
+
     self.panel = CreateFrame("Frame")
     self.panel.name = "Crit Wow"
+    self.db = CritWowDB
 
     local title = self.panel:CreateFontString("ARTWORK", nil, "GameFontNormalLarge")
     title:SetPoint("TOP")
     title:SetText("Crit Wow")
 
+    -- Play sounds check
     local cb = CreateFrame("CheckButton", nil, self.panel, "InterfaceOptionsCheckButtonTemplate")
     cb:SetPoint("TOPLEFT", 20, -20)
     cb.Text:SetText("Play sounds")
@@ -54,6 +94,7 @@ function f:InitializeOptions()
     end)
     cb:SetChecked(CritWowDB.playsSound)
 
+    -- Debounce slider
     local debounceSlider = CreateFrame("Slider", "CritWowDebounceSpeedSlider", self.panel, "OptionsSliderTemplate")
     debounceSlider:SetPoint("TOPLEFT", 20, -70)
     debounceSlider:SetMinMaxValues(0, 30)
@@ -71,6 +112,7 @@ function f:InitializeOptions()
         CritWow:SetDebounceSliderText()
     end)
 
+    -- Play test sound button
     local btn = CreateFrame("Button", nil, self.panel, "UIPanelButtonTemplate")
     btn:SetPoint("TOPLEFT", cb, 200, -10)
     btn:SetText("Play test sound")
@@ -79,6 +121,33 @@ function f:InitializeOptions()
     btn:SetScript("OnClick", function()
         CritWow:PlaySound()
     end)
+
+    -- Sound mode dropdown
+    local dropDown = CreateFrame("FRAME", nil, self.panel, "UIDropDownMenuTemplate")
+    dropDown:SetPoint("TOPLEFT", btn, -20, -30)
+    UIDropDownMenu_SetWidth(dropDown, 200)
+    UIDropDownMenu_SetText(dropDown, "Sound mode: " .. CritWow.modes[self.db.mode].name)
+
+    UIDropDownMenu_Initialize(dropDown, function(self, level, menuList)
+        local selectOption = UIDropDownMenu_CreateInfo()
+        selectOption.text = CritWow.modes.wowenWilson.name
+        selectOption.arg1 = CritWow.modes.wowenWilson.id
+        selectOption.checked = CritWowDB.mode == CritWow.modes.wowenWilson.id
+        selectOption.func = self.SetValue
+        UIDropDownMenu_AddButton(selectOption)
+
+        selectOption.text = CritWow.modes.dnc.name
+        selectOption.arg1 = CritWow.modes.dnc.id
+        selectOption.checked = CritWowDB.mode == CritWow.modes.dnc.id
+        UIDropDownMenu_AddButton(selectOption)
+
+    end)
+
+    function dropDown:SetValue(newValue)
+        CritWowDB.mode = newValue
+        UIDropDownMenu_SetText(dropDown, "Sound mode: " .. CritWow.modes[CritWowDB.mode].name)
+        CloseDropDownMenus()
+    end
 
     InterfaceOptions_AddCategory(self.panel)
 end